refactor: move the generic request and queryset helpers to the platform
`n23/core/utils.py` was a mixed bag: three request/redirect validators and two
model-agnostic queryset helpers sat alongside four functions that only make
sense for a gang. The generic five are used by platform code too, so the
edition was the wrong home — and `csrf_failure` could not move to the platform
while `safe_redirect` lived here.

Same split already applied to `gyrinx/models.py` / `n23/models.py`:

- `gyrinx/http.py` — `safe_redirect`, `build_safe_url`, `get_return_url`. All
  three validate a user-supplied URL against the requesting host, so they belong
  together and away from anything that might tempt a caller to skip them.
- `gyrinx/querysets.py` — `search_queryset`, `toggle_membership`.
- `n23/core/utils.py` keeps the four `get_list_*` helpers.

No behaviour change: the function bodies are moved verbatim. 28 files had their
imports rewritten, and the `search_queryset` tests move to
`gyrinx/tests/test_querysets.py` alongside the code they cover (11 tests, still
collected).

Import ordering on the touched files was fixed with a one-off
`ruff check --select I --fix`; the project's pinned rule selection is unchanged.
