+++ n23/core/utils.py
@@ -1,177 +1,11 @@
-"""
-Utility functions for the core app.
+"""Edition query helpers for lists.
+
+Generic helpers that used to live here moved to the platform: request/redirect
+validation to ``gyrinx.http``, and the model-agnostic queryset helpers to
+``gyrinx.querysets``. What remains knows about gangs.
 """
 
-from django.contrib.postgres.search import SearchQuery, SearchVector
 from django.db.models import Q
-from django.http import HttpResponseRedirect
-from django.utils.http import url_has_allowed_host_and_scheme
-
-
-def search_queryset(queryset, query, fields):
-    """
-    Apply a combined full-text and substring search to a queryset.
-
-    Uses PostgreSQL full-text search (SearchVector + SearchQuery) for
-    word-level matching, combined with icontains fallback on every field
-    so that partial/substring queries (e.g. "scav" matching "scavvies")
-    also return results.
-
-    Args:
-        queryset: The Django queryset to filter.
-        query: The search string. Stripped internally; if empty/None after
-            stripping, the queryset is returned unchanged.
-        fields: An iterable of field lookup strings to search across
-            (e.g. ["name", "summary", "owner__username"]).
-
-    Returns:
-        The filtered queryset, deduplicated via a PK subquery so that
-        reverse-FK and M2M search fields don't produce duplicate rows.
-
-    Example::
-
-        qs = search_queryset(
-            Pack.objects.all(),
-            request.GET.get("q", "").strip(),
-            ["name", "summary", "owner__username"],
-        )
-    """
-    query = (query or "").strip()
-    if not query:
-        return queryset
-
-    if not fields:
-        raise ValueError("search_queryset() requires at least one field")
-
-    # Build icontains fallback: OR across all fields
-    icontains_q = Q()
-    for field in fields:
-        icontains_q |= Q(**{f"{field}__icontains": query})
-
-    # Full-text search via SearchVector + SearchQuery
-    search_vector = SearchVector(*fields)
-    search_q = SearchQuery(query)
-
-    # Use a subquery to find matching PKs, then filter the original queryset.
-    # This avoids duplicates from JOINs created by SearchVector on related
-    # fields (e.g. reverse FK or M2M lookups like contentweaponprofile__name
-    # or contentweaponprofile__traits__name). With the annotation approach,
-    # DISTINCT is ineffective because the tsvector varies per joined row.
-    # Clear any inherited ordering so DISTINCT applies only to the PK column.
-    matching_pks = (
-        queryset.annotate(search=search_vector)
-        .filter(Q(search=search_q) | icontains_q)
-        .order_by()
-        .values_list("pk", flat=True)
-        .distinct()
-    )
-    return queryset.filter(pk__in=matching_pks)
-
-
-def safe_redirect(request, url, fallback_url="/"):
-    """
-    Perform a safe redirect, ensuring the URL is allowed.
-
-    Args:
-        request: The current HTTP request
-        url: The URL to redirect to
-        fallback_url: The URL to use if validation fails (default: "/")
-
-    Returns:
-        HttpResponseRedirect to either the validated URL or the fallback
-    """
-    if not url:
-        return HttpResponseRedirect(fallback_url)
-
-    if not url_has_allowed_host_and_scheme(
-        url, allowed_hosts={request.get_host()}, require_https=request.is_secure()
-    ):
-        return HttpResponseRedirect(fallback_url)
-
-    return HttpResponseRedirect(url)
-
-
-def toggle_membership(relation, user):
-    """
-    Toggle a user's membership in a many-to-many relation.
-
-    Used by the pin/star toggle views on lists and campaigns.
-
-    Returns:
-        True if the user is now a member, False if they were removed.
-    """
-    if relation.filter(pk=user.pk).exists():
-        relation.remove(user)
-        return False
-    relation.add(user)
-    return True
-
-
-def build_safe_url(request, path=None, query_string=None):
-    """
-    Build a safe URL from path and query string components.
-
-    Args:
-        request: The current HTTP request
-        path: The path component (default: request.path)
-        query_string: The query string (without '?')
-
-    Returns:
-        A safe URL string that can be used for redirects
-    """
-    # Use current path if not provided
-    if path is None:
-        path = request.path
-
-    # Build the full URL
-    if query_string:
-        url = f"{path}?{query_string}"
-    else:
-        url = path
-
-    # Validate the URL is safe
-    if not url_has_allowed_host_and_scheme(
-        url, allowed_hosts={request.get_host()}, require_https=request.is_secure()
-    ):
-        return path
-
-    return url
-
-
-def get_return_url(request, default_url):
-    """
-    Get a validated return URL from request parameters.
-
-    Extracts return_url from POST data (for form submissions) or GET parameters.
-    Validates the URL for security and falls back to default_url if invalid.
-
-    Args:
-        request: The HTTP request object
-        default_url: Fallback URL if return_url is missing or invalid
-
-    Returns:
-        str: A validated URL safe for redirects
-
-    Example:
-        default_url = reverse("core:list", args=(list.id,))
-        return_url = get_return_url(request, default_url)
-    """
-    # Check POST first (form submissions), then GET (query params)
-    return_url = request.POST.get("return_url") or request.GET.get(
-        "return_url", default_url
-    )
-
-    if not return_url:
-        return default_url
-
-    if not url_has_allowed_host_and_scheme(
-        return_url,
-        allowed_hosts={request.get_host()},
-        require_https=request.is_secure(),
-    ):
-        return default_url
-
-    return return_url
 
 
 def get_list_attributes(list_obj):
