+++ n23/core/views/campaign/assets.py
@@ -9,6 +9,9 @@ from django.shortcuts import get_object_or_404, redirect, render
 from django.urls import reverse
 
 from gyrinx import messages
+from gyrinx.analytics.models import EventNoun, EventVerb, log_event
+from gyrinx.http import get_return_url, safe_redirect
+from gyrinx.tracker import track
 from n23.core.forms.campaign import (
     AssetTransferForm,
     CampaignAssetCloneForm,
@@ -22,9 +25,6 @@ from n23.core.models.campaign import (
     CampaignAssetType,
     CampaignSubAsset,
 )
-from gyrinx.analytics.models import EventNoun, EventVerb, log_event
-from n23.core.utils import get_return_url, safe_redirect
-from gyrinx.tracker import track
 from n23.core.views.campaign.common import get_campaign_admin_or_404
