+++ n23/core/views/campaign/attributes.py
@@ -11,6 +11,8 @@ from django.shortcuts import get_object_or_404, redirect, render
 from django.urls import reverse
 
 from gyrinx import messages
+from gyrinx.analytics.models import EventNoun, EventVerb, log_event
+from gyrinx.http import get_return_url, safe_redirect
 from n23.core.forms.campaign import (
     CampaignAttributeTypeForm,
     CampaignAttributeValueForm,
@@ -21,8 +23,6 @@ from n23.core.models.campaign import (
     CampaignAttributeType,
     CampaignAttributeValue,
 )
-from gyrinx.analytics.models import EventNoun, EventVerb, log_event
-from n23.core.utils import get_return_url, safe_redirect
 from n23.core.views.campaign.common import get_campaign_admin_or_404
 
 logger = logging.getLogger(__name__)
