+++ n23/core/views/campaign/resources.py
@@ -10,20 +10,20 @@ from django.shortcuts import get_object_or_404, redirect, render
 from django.urls import reverse
 
 from gyrinx import messages
+from gyrinx.analytics.models import EventNoun, EventVerb, log_event
+from gyrinx.http import get_return_url, safe_redirect
+from gyrinx.tracker import track
 from n23.core.forms.campaign import CampaignResourceTypeForm, ResourceModifyForm
 from n23.core.models.campaign import (
     Campaign,
     CampaignListResource,
     CampaignResourceType,
 )
-from gyrinx.analytics.models import EventNoun, EventVerb, log_event
-from n23.core.utils import get_return_url, safe_redirect
 from n23.core.views.campaign.common import (
     ensure_campaign_list_resources,
+    get_campaign_admin_or_404,
     get_campaign_resource_types_with_resources,
 )
-from gyrinx.tracker import track
-from n23.core.views.campaign.common import get_campaign_admin_or_404
 
 # Constants for transaction limits
 MAX_CREDITS = 10000
