+++ n23/core/views/csrf.py
@@ -3,7 +3,7 @@ from django.urls import reverse
 from django.views.decorators.csrf import csrf_exempt
 
 from gyrinx.analytics.models import EventNoun, EventVerb, log_event
-from n23.core.utils import safe_redirect
+from gyrinx.http import safe_redirect
 
 
 @csrf_exempt
