+++ n23/core/views/fighter/equipment_sets.py
@@ -15,8 +15,8 @@ from django.urls import reverse
 
 from gyrinx import messages
 from gyrinx.analytics.models import EventNoun, EventVerb, log_event
+from gyrinx.http import safe_redirect
 from n23.core.models.list import ListFighterEquipmentSet
-from n23.core.utils import safe_redirect
 from n23.core.views.fighter.permissions import get_list_and_fighter
