+++ n23/core/views/fighter/narrative.py
@@ -4,13 +4,13 @@ from django.contrib.auth.decorators import login_required
 from django.shortcuts import get_object_or_404, render
 from django.urls import reverse
 
+from gyrinx.analytics.models import EventNoun, EventVerb, log_event
+from gyrinx.http import get_return_url, safe_redirect
 from n23.core.forms.list import (
     EditListFighterNarrativeForm,
     EditListFighterNotesForm,
 )
-from gyrinx.analytics.models import EventNoun, EventVerb, log_event
 from n23.core.models.list import List, ListFighter
-from n23.core.utils import get_return_url, safe_redirect
 from n23.core.views.list.common import get_clean_list_or_404
