+++ n23/core/views/list/views.py
@@ -15,28 +15,25 @@ from django.views import generic
 from django.views.decorators.http import require_POST
 
 from gyrinx import messages
-from n23.content.models import ContentEquipment, ContentFighter, ContentHouse
+from gyrinx.analytics.models import EventNoun, EventVerb, log_event
+from gyrinx.http import build_safe_url, get_return_url, safe_redirect
+from gyrinx.models import is_valid_uuid
+from gyrinx.querysets import search_queryset, toggle_membership
 from gyrinx.site.models import BANNER_CACHE_KEY
+from gyrinx.tracing import traced
+from gyrinx.tracker import track
+from n23.content.models import ContentEquipment, ContentFighter, ContentHouse
 from n23.core.forms.list import CloneListForm, EditListForm, NewListForm
 from n23.core.handlers.list import handle_list_clone, handle_list_creation
 from n23.core.models.list import List, ListFighter
-from gyrinx.analytics.models import EventNoun, EventVerb, log_event
 from n23.core.utils import (
-    build_safe_url,
     get_list_attributes,
     get_list_campaign_resources,
     get_list_held_assets,
     get_list_recent_campaign_actions,
-    get_return_url,
-    safe_redirect,
-    search_queryset,
-    toggle_membership,
 )
-from n23.core.views.list.common import get_clean_list_or_404
-from gyrinx.models import is_valid_uuid
-from gyrinx.tracing import traced
-from gyrinx.tracker import track
 from n23.core.views.fighter.permissions import arbitrator_q
+from n23.core.views.list.common import get_clean_list_or_404
 
 
 class ListsListView(generic.ListView):
