+++ n23/core/views/notification.py
@@ -20,8 +20,8 @@ from django.utils import timezone
 from django.views import generic
 from django.views.decorators.http import require_POST
 
+from gyrinx.http import safe_redirect
 from n23.core.models.notification import Notification, NotificationType
-from n23.core.utils import safe_redirect
 
 VALID_BUCKETS = {"inbox", "archived"}
 VALID_STATUSES = {"all", "unread", "read"}
