+++ n23/core/views/pack.py
@@ -21,6 +21,10 @@ from django.views import generic
 from pydantic import BaseModel, ValidationError
 
 from gyrinx import messages
+from gyrinx.analytics.models import EventNoun, EventVerb, log_event
+from gyrinx.http import safe_redirect
+from gyrinx.models import is_valid_uuid
+from gyrinx.querysets import search_queryset
 from n23.content.models.attribute import ContentAttribute, ContentAttributeValue
 from n23.content.models.default_assignment import ContentFighterDefaultAssignment
 from n23.content.statlines import (
@@ -43,15 +47,15 @@ from n23.content.models.equipment_list import (
 from n23.content.models.fighter import ContentFighter
 from n23.content.models.house import ContentHouse
 from n23.content.models.metadata import ContentRule
-from n23.content.models.skill import ContentSkill, ContentSkillCategory
-from n23.content.models.statline import (
-    ContentStatlineType,
-)
 from n23.content.models.psyker import (
     ContentFighterPsykerDisciplineAssignment,
     ContentFighterPsykerPowerDefaultAssignment,
     ContentPsykerDiscipline,
     ContentPsykerPower,
+)
+from n23.content.models.skill import ContentSkill, ContentSkillCategory
+from n23.content.models.statline import (
+    ContentStatlineType,
 )
 from n23.content.models.weapon import (
     ContentWeaponAccessory,
@@ -59,6 +63,7 @@ from n23.content.models.weapon import (
     ContentWeaponTrait,
 )
 from n23.core.forms.pack import (
+    HOUSE_RULE_TARGET_CHOICES,
     ContentAttributePackForm,
     ContentAttributeValuePackForm,
     ContentFighterPackForm,
@@ -77,10 +82,8 @@ from n23.core.forms.pack import (
     EquipmentModifiersForm,
     PackAttachmentForm,
     PackForm,
-    HOUSE_RULE_TARGET_CHOICES,
 )
 from n23.core.models.campaign import Campaign, CampaignContentPack
-from gyrinx.analytics.models import EventNoun, EventVerb, log_event
 from n23.core.models.list import List
 from n23.core.models.pack import (
     PACK_ATTACHMENT_MAX_PER_PACK,
@@ -89,8 +92,6 @@ from n23.core.models.pack import (
     CustomContentPackItem,
     CustomContentPackPermission,
 )
-from n23.core.utils import safe_redirect, search_queryset
-from gyrinx.models import is_valid_uuid
 from n23.models import FighterCategoryChoices
